First stage of cleaning code

+++ deps/md5/md5.c
@@ -288,7 +288,7 @@ void MD5_Final(unsigned char *result, MD5_CTX *ctx)
 	memset(ctx, 0, sizeof(*ctx));
 }
 
-unsigned char *MD5(void *pointer, size_t pointerLength , unsigned char *md5HashPointer)
+unsigned char *MD5(void *pointer, unsigned long pointerLength , unsigned char *md5HashPointer)
 {
 
     MD5_CTX mdContext;
